feat(mcp): move MCP package out of tools with backwards-compat aliases
Promote the MCP integration from src/strands/tools/mcp to a top-level
src/strands/mcp package. MCP now spans tools, prompts, resources, tasks,
and elicitation, so it no longer belongs under tools/.

Backwards compatibility is preserved: strands.tools.mcp and its submodules
continue to work, re-export the canonical classes (object identity held),
and emit DeprecationWarning pointing users to the new import path. A
dedicated test_deprecated_aliases.py suite locks the contract down.

Closes #1431
