+++ tests/strands/mcp/conftest.py
@@ -37,7 +37,7 @@ def mock_session():
     mock_session_cm.__aenter__.return_value = mock_session
 
     # Patch ClientSession to return our mock session
-    with patch("strands.tools.mcp.mcp_client.ClientSession", return_value=mock_session_cm):
+    with patch("strands.mcp.mcp_client.ClientSession", return_value=mock_session_cm):
         yield mock_session
