+++ tests/strands/tools/test_registry.py
@@ -7,9 +7,9 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 import strands
+from strands.mcp import MCPClient
 from strands.tools import PythonAgentTool, ToolProvider
 from strands.tools.decorator import DecoratedFunctionTool, tool
-from strands.tools.mcp import MCPClient
 from strands.tools.registry import ToolRegistry
